[PLAYER-5447] Refactor views components to use own styles
