+++ sdk/react/src/views/AdPlaybackScreen/AdPlaybackScreen.styles.js
@@ -1,0 +1,17 @@
+// @flow
+
+import { StyleSheet } from 'react-native';
+
+export default StyleSheet.create({
+  adContainer: {
+    flex: 1,
+    flexDirection: 'column',
+    backgroundColor: 'transparent',
+    overflow: 'hidden',
+  },
+  placeholder: {
+    flex: 1,
+    alignItems: 'stretch',
+    backgroundColor: 'transparent',
+  },
+});
