+++ sdk/react/src/views/CastConnectedScreen/CastPlayPauseButtons/CastPlayPauseButtons.js
@@ -11,8 +11,7 @@ import * as Utils from '../../../lib/utils';
 import SkipButton from '../../../shared/SkipButton';
 import SwitchButton from './SwitchButton';
 
-// TODO(styles): Prefer own styles instead.
-import styles from '../../../shared/styles/rectangularButton.styles';
+import styles from './CastPlayPauseButtons.styles';
 
 const PLAY = 'play';
 const PAUSE = 'pause';
