+++ sdk/react/src/views/CastConnectedScreen/CastPlayPauseButtons/CastPlayPauseButtons.styles.js
@@ -1,0 +1,9 @@
+// @flow
+
+import { StyleSheet } from 'react-native';
+
+export default StyleSheet.create({
+  buttonTextStyle: {
+    color: 'white',
+  },
+});
