+++ sdk/react/src/views/CastDevicesScreen/CastDevicesScreen.styles.js
@@ -12,27 +12,6 @@ export default StyleSheet.create({
     marginBottom: 25,
     flex: 0,
   },
-  itemContainer: {
-    flex: 1,
-    padding: 12,
-    flexDirection: 'row',
-    alignItems: 'center',
-  },
-  itemContainerSelected: {
-    flex: 1,
-    backgroundColor: 'rgba(114, 114, 114, 0.6)',
-    padding: 12,
-    flexDirection: 'row',
-    alignItems: 'center',
-  },
-  scrollContainer: {
-    flex: 1,
-    justifyContent: 'center',
-    flexDirection: 'column',
-    alignItems: 'center',
-    backgroundColor: 'transparent',
-    marginTop: 30,
-  },
   dismissButtonTopRight: {
     position: 'absolute',
     top: 0,
@@ -40,22 +19,6 @@ export default StyleSheet.create({
   },
   iconDismiss: {
     padding: 25,
-  },
-  icon: {
-    paddingLeft: 30,
-  },
-  text: {
-    paddingLeft: 36,
-    paddingRight: 100,
-    color: '#FFFFFF',
-    fontFamily: 'AvenirNext-DemiBold',
-    fontSize: 16,
-  },
-  textSelected: {
-    paddingLeft: 36,
-    paddingRight: 100,
-    fontFamily: 'AvenirNext-DemiBold',
-    fontSize: 16,
   },
   title: {
     position: 'absolute',
