+++ sdk/react/src/views/DiscoveryPanel/DiscoveryPanel.js
@@ -11,8 +11,6 @@ import ResponsiveList from './ResponsiveList';
 import CountdownViewAndroid from '../../shared/CountdownTimerAndroid';
 import CountdownViewiOS from '../../shared/CountdownTimerIos';
 
-// TODO(styles): Prefer own styles instead.
-import panelStyles from '../styles/view.styles';
 import styles from './DiscoveryPanel.styles';
 
 // TODO: read this from config.
@@ -278,24 +276,24 @@ Regular CountdownView uses onTimerCompleted callback defined in jsx
     // This ability is added in latest react native 0.46 onwards
     // so we can remove this piece of code once we upgrade.
     return (
-      <View style={panelStyles.panelTitleView}>
-        <Text style={[panelStyles.panelTitleText]}>
+      <View style={styles.panelTitleView}>
+        <Text style={[styles.panelTitleText]}>
           {title}
         </Text>
         <TouchableHighlight accessible accessibilityLabel={BUTTON_NAMES.DISCOVERY}>
           <View>
-            <Text style={panelStyles.panelIcon}>{panelIcon}</Text>
+            <Text style={styles.panelIcon}>{panelIcon}</Text>
           </View>
         </TouchableHighlight>
-        <View style={panelStyles.headerFlexibleSpace} />
+        <View style={styles.headerFlexibleSpace} />
         <TouchableHighlight
           accessible
           accessibilityLabel={BUTTON_NAMES.DISMISS}
           accessibilityComponentType="button"
-          style={[panelStyles.dismissButton]}
+          style={[styles.dismissButton]}
           onPress={this.onDismissPress}
         >
-          <Text style={panelStyles.dismissIcon}>{icons.dismiss.fontString}</Text>
+          <Text style={styles.dismissIcon}>{icons.dismiss.fontString}</Text>
         </TouchableHighlight>
       </View>
     );
@@ -344,7 +342,7 @@ Regular CountdownView uses onTimerCompleted callback defined in jsx
     }
     const animationStyle = { opacity };
     return (
-      <Animated.View style={[panelStyles.panel, animationStyle]}>
+      <Animated.View style={[styles.panel, animationStyle]}>
         {this.renderHeader()}
         {this.renderList(itemRect, thumbnailStyle, columnContainerStyle)}
         {this.renderError()}
