+++ sdk/react/src/views/VideoView/VideoView.styles.js
@@ -3,19 +3,8 @@
 import { StyleSheet } from 'react-native';
 
 export default StyleSheet.create({
-  fullscreenContainer: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
   container: {
     flex: 0,
-    flexDirection: 'column',
-    backgroundColor: 'transparent',
-    overflow: 'hidden',
-  },
-  adContainer: {
-    flex: 1,
     flexDirection: 'column',
     backgroundColor: 'transparent',
     overflow: 'hidden',
@@ -24,5 +13,31 @@ export default StyleSheet.create({
     flex: 1,
     alignItems: 'stretch',
     backgroundColor: 'transparent',
+  },
+  closedCaptionsContainer: {
+    backgroundColor: 'transparent',
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+  closedCaptions: {
+    alignSelf: 'center',
+    textAlign: 'center',
+    padding: 4,
+    fontWeight: 'bold',
+    fontSize: 16,
+    fontFamily: 'Helvetica',
+    color: 'white',
+    backgroundColor: 'rgba(0, 0, 0, 0.7)',
+  },
+  dismissOverlay: {
+    position: 'absolute',
+    left: 10,
+    top: 10,
+  },
+  dismissIcon: {
+    flex: 1,
+    fontFamily: 'ooyala-slick-type',
+    fontSize: 20,
+    color: 'white',
   },
 });
